+++ x/token/handler_test.go
@@ -5,9 +5,10 @@ import (
 	"fmt"
 	"testing"
 
+	"github.com/medibloc/panacea-core/x/token/internal"
+
 	"github.com/medibloc/panacea-core/x/token/types"
 
-	"github.com/cosmos/cosmos-sdk/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/stretchr/testify/require"
 )
@@ -19,7 +20,7 @@ var (
 
 func TestHandleMsgIssueToken(t *testing.T) {
 	ctx := sdk.Context{}.WithTxBytes(txBytes)
-	keeper := newMockKeeper()
+	keeper := internal.NewMockKeeper()
 	msg := types.NewMsgIssueToken("my token", "LOV", sdk.NewInt(1000000000), true, sdk.AccAddress{})
 
 	res := handleMsgIssueToken(ctx, keeper, msg)
@@ -40,7 +41,7 @@ func TestHandleMsgIssueToken(t *testing.T) {
 
 func TestHandleMsgIssueToken_Exists(t *testing.T) {
 	ctx := sdk.Context{}.WithTxBytes(txBytes)
-	keeper := newMockKeeper()
+	keeper := internal.NewMockKeeper()
 	msg := types.NewMsgIssueToken("my token", "LOV", sdk.NewInt(1000000000), true, sdk.AccAddress{})
 
 	res := handleMsgIssueToken(ctx, keeper, msg)
@@ -51,33 +52,4 @@ func TestHandleMsgIssueToken_Exists(t *testing.T) {
 	require.False(t, res.IsOK())
 	require.Equal(t, types.DefaultCodespace, res.Codespace)
 	require.Equal(t, types.CodeTokenExists, res.Code)
-}
-
-// mockKeeper implements the token.Keeper interface
-type mockKeeper struct {
-	tokens map[types.Symbol]types.Token
-}
-
-func newMockKeeper() *mockKeeper {
-	return &mockKeeper{tokens: make(map[types.Symbol]types.Token)}
-}
-
-func (k mockKeeper) Codec() *codec.Codec {
-	return codec.New()
-}
-
-func (k mockKeeper) SetToken(ctx sdk.Context, symbol types.Symbol, token types.Token) {
-	k.tokens[symbol] = token
-}
-
-func (k mockKeeper) GetToken(ctx sdk.Context, symbol types.Symbol) types.Token {
-	return k.tokens[symbol]
-}
-
-func (k mockKeeper) ListTokens(ctx sdk.Context) []types.Symbol {
-	symbols := make([]types.Symbol, 0)
-	for symbol := range k.tokens {
-		symbols = append(symbols, symbol)
-	}
-	return symbols
 }
